Made the pathfinding look a little less like a drunkard's walk

+++ src/dungeon.rs
@@ -88,7 +88,7 @@ fn pick_room() -> (Vec<Vec<Tile>>, usize, usize) {
         // Now turn all the squares inside the circle into floors
         for r in 1..height - 1 {
             for c in 1..width - 1 {
-                if util::distance(r as i32, c as i32, rc as i32, cc as i32) <= radius as i32 {
+                if util::distance(r as i32, c as i32, rc as i32, cc as i32) as usize <= radius {
                     room[r][c] = Tile::StoneFloor;
                 }                
             }
